refactor : move functions to utils
